Agregué ProductoService, ProveedorService y EnvioService. Y ademas, coregí Repositories

+++ src/main/java/com/duoc/SpringApp_Grupo5/Repositorio/LogisticayEnvio/EnvioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EnvioRepository extends JpaRepository<Envio, Long> {
+public interface EnvioRepository extends JpaRepository<Envio, Integer> {
 }
